Split services and controllers into seperate files
