+++ package.json
@@ -32,6 +32,7 @@
         "gulp-util": "^2.2.14",
         "moment": "^2.14.1",
         "ngstorage": "^0.3.11",
+        "require-globify": "^1.4.1",
         "shelljs": "^0.3.0",
         "vinyl-source-stream": "^1.1.0"
     },
